+++ Sources/PippinCore/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- A secondary foreground color in `Colors` (and a default value in `DefaultColors`).
+
 ## [2.1.0] 2020-08-08
 
 ### Added